requirements.txt
accelerate==1.12.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-cors==0.8.1
aiosignal==1.4.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.0
appdirs==1.4.4
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
backoff==2.2.1
bce-python-sdk==0.9.59
beartype==0.15.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
bokeh==3.8.1
boto3==1.42.21
botocore==1.42.21
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
deltalake==0.17.4
Deprecated==1.3.1
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
docling==2.66.0
docling-core==2.57.0
docling-ibm-models==3.10.3
docling-parse==4.7.2
effdet==0.4.1
einops==0.8.1
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.1
Faker==40.1.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.2
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fs==2.4.16
fsspec==2025.12.0
ftfy==6.3.1
future==1.0.0
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.5.0
google-cloud-pubsub==2.34.0
google-cloud-vision==3.11.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h3==4.4.1
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.1
ipython==9.9.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_bokeh==4.0.5
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
langdetect==1.0.9
latex2mathml==3.78.1
linkify-it-py==2.0.3
lxml==6.0.2
lxml_html_clean==0.4.3
Markdown==3.10
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
ml_dtypes==0.5.4
modelscope==1.33.0
more-itertools==10.8.0
mpire==2.10.2
mpmath==1.3.0
msoffcrypto-tool==5.4.2
multidict==6.7.0
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.14.0
networkx==3.6.1
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
olefile==0.47
omegaconf==2.3.0
onnx==1.20.0
onnxruntime==1.23.2
openai==2.14.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==25.0
paddleocr==3.3.2
paddlex==3.3.12
pandas==2.3.3
panel==1.8.5
param==2.3.1
parso==0.8.5
pathway==0.27.1
pdf2image==1.17.0
pdfminer.six==20250506
pexpect==4.9.0
pi_heif==1.1.1
pikepdf==10.1.0
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.2.0
premailer==3.10.0
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycocotools==2.0.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
pypandoc==1.16.2
pyparsing==3.3.1
pypdf==6.5.0
pypdfium2==4.30.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-iso639==2025.11.16
python-magic==0.4.27
python-multipart==0.0.21
python-oxmsg==0.0.2
python-pptx==1.0.2
python-sat==1.8.dev26
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.2
RapidFuzz==3.14.3
rapidocr==3.4.5
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
ruamel.yaml==0.19.1
s3transfer==0.16.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
semchunk==2.2.2
sentence-transformers==5.2.0
sentencepiece==0.2.1
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.1
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.22
tinysegmenter==0.3
tldextract==5.3.1
tokenizers==0.22.2
torch==2.9.1
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-java==0.23.5
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
triton==3.5.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
ujson==5.11.0
unstructured==0.18.24
unstructured-client==0.42.6
unstructured.pytesseract==0.3.15
unstructured_inference==1.1.4
uritemplate==4.2.0
urllib3==2.6.2
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.1.4
widgetsnbextension==4.0.15
wrapt==2.0.1
xlrd==2.0.2
xlsxwriter==3.2.9
xyzservices==2025.11.0
yarl==1.22.0
zipp==3.23.0
